# docs/swagger.yaml
basePath: /
definitions:
  dto.CollectibleListResponseDTO:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.CollectibleResponseDTO'
        type: array
      meta:
        $ref: '#/definitions/dto.PaginationMetaDTO'
    type: object
  dto.CollectibleResponseDTO:
    properties:
      condition:
        example: excellent
        type: string
      country:
        example: СССР
        type: string
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      name:
        example: Монета 1 рубль 1961 года
        type: string
      price:
        example: 5000
        type: number
      updated_at:
        example: "2024-01-02T00:00:00Z"
        type: string
      user_id:
        example: 661e8400-e29b-41d4-a716-446655440001
        type: string
      year:
        example: 1961
        type: integer
    type: object
  dto.CreateCollectibleDTO:
    properties:
      condition:
        example: excellent
        type: string
      country:
        example: СССР
        type: string
      name:
        example: Монета 1 рубль 1961 года
        type: string
      price:
        example: 5000
        type: number
      year:
        example: 1961
        type: integer
    required:
    - country
    - name
    - price
    - year
    type: object
  dto.ErrorDTO:
    properties:
      error:
        example: Something went wrong
        type: string
    type: object
...
  dto.MessageDTO:
    properties:
      message:
        example: Operation successful
        type: string
    type: object
  dto.PaginationMetaDTO:
    properties:
      limit:
        example: 10
        type: integer
      page:
        example: 1
        type: integer
      total:
        example: 100
        type: integer
      totalPages:
        example: 10
        type: integer
    type: object
  dto.PatchCollectibleDTO:
    properties:
      condition:
        example: excellent
        type: string
      country:
        example: СССР
        type: string
      name:
        example: Монета 1 рубль 1961 года
        type: string
      price:
        example: 5000
        type: number
      year:
        example: 1961
        type: integer
    type: object
...
  dto.UpdateCollectibleDTO:
    properties:
      condition:
        example: excellent
        type: string
      country:
        example: СССР
        type: string
      name:
        example: Монета 1 рубль 1961 года
        type: string
      price:
        example: 5000
        type: number
      year:
        example: 1961
        type: integer
    type: object
...
host: localhost:4200
info:
  contact: {}
  description: REST API для управления коллекционными предметами с JWT авторизацией,
    OAuth 2.0, MinIO и Swagger.
  title: NewYear API
  version: "1.0"
paths:
...
  /files:
    post:
      consumes:
      - multipart/form-data
      description: Загружает файл в MinIO. Поддерживает PNG и JPEG до 10MB
      parameters:
      - description: Файл для загрузки
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Загрузить файл
      tags:
      - Files
  /files/{fileId}:
    delete:
      description: Удаляет файл из MinIO и БД
      parameters:
      - description: ID файла
        in: path
        name: fileId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Удалить файл
      tags:
      - Files
    get:
      description: Скачивает файл из MinIO по ID
      parameters:
      - description: ID файла
        in: path
        name: fileId
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: Файл
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Скачать файл
      tags:
      - Files
  /items:
    get:
      description: Возвращает список коллекционных предметов текущего пользователя
        с пагинацией.
      parameters:
      - default: 1
        description: Номер страницы (по умолчанию 1)
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 10
        description: Предметов на странице (по умолчанию 10, максимум 100)
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CollectibleListResponseDTO'
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/dto.ErrorDTO'
        "500":
          description: Внутренняя ошибка
          schema:
            $ref: '#/definitions/dto.ErrorDTO'
      security:
      - CookieAuth: []
      summary: Получить список предметов
      tags:
      - Items
    post:
      consumes:
      - application/json
      description: Создаёт новый предмет и привязывает его к авторизованному пользователю.
      parameters:
      - description: Данные предмета
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateCollectibleDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.MessageDTO'
        "400":
          description: Неверный формат данных
          schema:
            $ref: '#/definitions/dto.ErrorDTO'
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/dto.ErrorDTO'
        "500":
          description: Внутренняя ошибка
          schema:
            $ref: '#/definitions/dto.ErrorDTO'
      security:
      - CookieAuth: []
      summary: Создать коллекционный предмет
      tags:
      - Items
  /items/{id}:
    delete:
      description: Мягкое удаление — запись остаётся в БД с заполненным полем deleted_at.
        Доступно только владельцу.
      parameters:
      - description: UUID предмета
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content — предмет удалён
          schema:
            type: string
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/dto.ErrorDTO'
        "403":
          description: Нет доступа к чужому предмету
          schema:
            $ref: '#/definitions/dto.ErrorDTO'
        "404":
          description: Предмет не найден
          schema:
            $ref: '#/definitions/dto.ErrorDTO'
      security:
      - CookieAuth: []
      summary: Удалить предмет (Soft Delete)
      tags:
      - Items
    get:
      description: Возвращает коллекционный предмет по UUID. Доступен только владельцу.
      parameters:
      - description: UUID предмета
        example: 550e8400-e29b-41d4-a716-446655440000
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CollectibleResponseDTO'
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/dto.ErrorDTO'
        "403":
          description: Нет доступа к чужому предмету
          schema:
            $ref: '#/definitions/dto.ErrorDTO'
        "404":
          description: Предмет не найден
          schema:
            $ref: '#/definitions/dto.ErrorDTO'
      security:
      - CookieAuth: []
      summary: Получить предмет по ID
      tags:
      - Items
    patch:
      consumes:
      - application/json
      description: Обновляет только переданные поля предмета. Доступно только владельцу.
      parameters:
      - description: UUID предмета
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Поля для обновления
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.PatchCollectibleDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CollectibleResponseDTO'
        "400":
          description: Неверный формат данных
          schema:
            $ref: '#/definitions/dto.ErrorDTO'
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/dto.ErrorDTO'
        "403":
          description: Нет доступа к чужому предмету
          schema:
            $ref: '#/definitions/dto.ErrorDTO'
        "404":
          description: Предмет не найден
          schema:
            $ref: '#/definitions/dto.ErrorDTO'
      security:
      - CookieAuth: []
      summary: Частичное обновление предмета
      tags:
      - Items
    put:
      consumes:
      - application/json
      description: Полностью заменяет данные предмета. Доступно только владельцу.
      parameters:
      - description: UUID предмета
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Новые данные предмета
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateCollectibleDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CollectibleResponseDTO'
        "400":
          description: Неверный формат данных
          schema:
            $ref: '#/definitions/dto.ErrorDTO'
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/dto.ErrorDTO'
        "403":
          description: Нет доступа к чужому предмету
          schema:
            $ref: '#/definitions/dto.ErrorDTO'
        "404":
          description: Предмет не найден
          schema:
            $ref: '#/definitions/dto.ErrorDTO'
      security:
      - CookieAuth: []
      summary: Полное обновление предмета
      tags:
      - Items
...
securityDefinitions:
  CookieAuth:
    description: JWT access token через HttpOnly cookie.
    in: cookie
    name: access_token
    type: apiKey
swagger: "2.0"